requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
coverage==7.6.1
django==4.2.23
django-coverage-plugin==3.1.0
djangorestframework==3.14.0
pytz==2025.2
sqlparse==0.5.3
typing-extensions==4.13.2
